修复符号饱和加法问题

+++ arm/src/inst/qadd8.c
@@ -21,14 +21,14 @@ int32 arm_inst_qadd8(cpu_state_t *st, uint32 inst)
 		return EXEC_UNPREDICTABLE;
 
 	//低位结果
-	int64 resLL = bitm(regv(rm), 0, 7) + bitm(regv(rn), 0, 7);
+	int16 resLL = (int16)(int8)bitm(regv(rm), 0, 7) + (int8)bitm(regv(rn), 0, 7);
 	uint32 LL = signed_sat(resLL, 8, &issat);
-	int64 resLH = bitm(regv(rm), 8, 15) + bitm(regv(rn), 8, 15);
+	int16 resLH = (int16)(int8)bitm(regv(rm), 8, 15) + (int8)bitm(regv(rn), 8, 15);
 	uint32 LH = signed_sat(resLH, 8, &issat);
 	//高位结果
-	int64 resHL = bitm(regv(rm), 16, 23) + bitm(regv(rn), 16, 23);
+	int16 resHL = (int16)(int8)bitm(regv(rm), 16, 23) + (int8)bitm(regv(rn), 16, 23);
 	uint32 HL = signed_sat(resHL, 8, &issat);
-	int64 resHH = bitm(regv(rm), 24, 31) + bitm(regv(rn), 24, 31);
+	int16 resHH = (int16)(int8)bitm(regv(rm), 24, 31) + (int8)bitm(regv(rn), 24, 31);
 	uint32 HH = signed_sat(resHH, 8, &issat);
 
 	//结果综合
