+++ arm/src/inst/qaddsubx.c
@@ -21,10 +21,10 @@ int32 arm_inst_qaddsubx(cpu_state_t *st, uint32 inst)
 		return EXEC_UNPREDICTABLE;
 
 	//低位结果
-	int64 resL = bitm(regv(rn), 0, 15) - bitm(regv(rm), 16, 31);
+	int32 resL = (int32)(int16)bitm(regv(rn), 0, 15) - (int32)(int16)bitm(regv(rm), 16, 31);
 	uint32 L = signed_sat(resL, 16, &issat);
 	//高位结果
-	int64 resH = bitm(regv(rn), 16, 31) + bitm(regv(rm), 0, 15);
+	int32 resH = (int32)(int16)bitm(regv(rn), 16, 31) + (int32)(int16)bitm(regv(rm), 0, 15);
 	uint32 H = signed_sat(resH, 16, &issat);
 
 	//结果综合
